requirements.txt
streamlit==1.32.1
python-dotenv==1.0.1
openai==1.63.1 